acme/requirements.txt
pandas
scikit-learn
tpot
PyYAML
certifi